make notification component
